+++ playbook/spec/pb_kits/playbook/kits/table_header_spec.rb
@@ -21,7 +21,7 @@ RSpec.describe Playbook::PbTable::TableHeader do
   }
   it {
     is_expected.to define_enum_prop(:placement)
-      .with_default("bottom-start")
+      .with_default("bottom-end")
       .with_values("top", "right", "bottom", "left", "top-start", "top-end", "bottom-start", "bottom-end", "right-start", "right-end", "left-start", "left-end")
   }
   it {
@@ -29,12 +29,40 @@ RSpec.describe Playbook::PbTable::TableHeader do
       .of_type(Playbook::Props::HashArray)
       .with_default([{}])
   }
+  it {
+    is_expected.to define_prop(:sort_dropdown)
+      .of_type(Playbook::Props::Boolean)
+      .with_default(false)
+  }
   it { is_expected.to define_prop(:text).with_default("") }
 
   describe "#classname" do
     it "returns namespaced class name", :aggregate_failures do
       expect(subject.new({}).classname).to start_with "pb_table_header_kit"
       expect(subject.new(dark: true, align: "end").classname).to eq "pb_table_header_kit_align_end dark align_content_center"
+    end
+  end
+
+  describe "next_link" do
+    it "makes the next link the first sort option when nothing is active" do
+      expect(
+        subject.new(
+          sort_menu: [
+            { item: "Territory", link: "?sort=territory_desc", active: false, direction: "desc" },
+            { item: "Territory", link: "?sort=territory_asc", active: false, direction: "asc" },
+          ]
+        ).next_link
+      ).to eq "?sort=territory_desc"
+    end
+    it "provides what the next sort option link would be" do
+      expect(
+        subject.new(
+          sort_menu: [
+            { item: "Territory", link: "?sort=territory_desc", active: true, direction: "desc" },
+            { item: "Territory", link: "?sort=territory_asc", active: false, direction: "asc" },
+          ]
+        ).next_link
+      ).to eq "?sort=territory_asc"
     end
   end
 end